Refactor another script
